refactored Alerts to separate component
