+++ grafana-plugin/src/PluginPage.tsx
@@ -3,21 +3,16 @@ import React from 'react';
 import { PluginPageProps, PluginPage as RealPluginPage } from '@grafana/runtime';
 import Header from 'navbar/Header/Header';
 
+import Alerts from 'components/Alerts/Alerts';
 import { pages } from 'pages';
 import { isTopNavbar } from 'plugin/GrafanaPluginRootPage.helpers';
 import { useStore } from 'state/useStore';
 import { DEFAULT_PAGE } from 'utils/consts';
 import { useQueryParams } from 'utils/hooks';
 
-export const PluginPage = (
-  isTopNavbar() ? RealPlugin : PluginPageFallback
-) as React.ComponentType<ExtendedPluginPageProps>;
+export const PluginPage = (isTopNavbar() ? RealPlugin : PluginPageFallback) as React.ComponentType<PluginPageProps>;
 
-interface ExtendedPluginPageProps extends PluginPageProps {
-  renderAlertsFn?: () => React.ReactNode;
-}
-
-function RealPlugin(props: ExtendedPluginPageProps): React.ReactNode {
+function RealPlugin(props: PluginPageProps): React.ReactNode {
   const store = useStore();
 
   const queryParams = useQueryParams();
@@ -26,7 +21,7 @@ function RealPlugin(props: ExtendedPluginPageProps): React.ReactNode {
   return (
     <RealPluginPage {...props}>
       {/* Render alerts at the top */}
-      {props.renderAlertsFn && props.renderAlertsFn()}
+      <Alerts />
       <Header page={page} backendLicense={store.backendLicense} />
       {pages[page].text && <h3 className="page-title">{pages[page].text}</h3>}
       {props.children}
@@ -34,6 +29,6 @@ function RealPlugin(props: ExtendedPluginPageProps): React.ReactNode {
   );
 }
 
-export function PluginPageFallback(props: ExtendedPluginPageProps): React.ReactNode {
+export function PluginPageFallback(props: PluginPageProps): React.ReactNode {
   return props.children;
 }
